blk-cgroup: Allow sleeping while dynamically allocating a group
Currently, all the cfq_group or throtl_group allocations happen while
we are holding ->queue_lock and sleeping is not allowed.

Soon, we will move to per cpu stats and also need to allocate the
per group stats. As one can not call alloc_percpu() from atomic
context as it can sleep, we need to drop ->queue_lock, allocate the
group, retake the lock and continue processing.

In throttling code, I check the queue DEAD flag again to make sure
that driver did not call blk_cleanup_queue() in the mean time.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>

Conflicts:

	block/blk-throttle.c
	block/cfq-iosched.c

Change-Id: I1f062c516882457a66059a63a553f4e55b514531
Reviewed-on: http://coconut.acer.com.tw:8080/10897
Reviewed-by: vincent_tang <vincent_tang@acer.com.tw>
Tested-by: vincent_tang <vincent_tang@acer.com.tw>
Reviewed-by: elay_hu <elay_hu@acer.com.tw>
Reviewed-on: http://coconut.acer.com.tw:8080/11809