block: remove support for bio remapping from ->make_request
There is very little benefit in allowing to let a ->make_request
instance update the bios device and sector and loop around it in
__generic_make_request when we can archive the same through calling
generic_make_request from the driver and letting the loop in
generic_make_request handle it.

Note that various drivers got the return value from ->make_request and
returned non-zero values for errors.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Acked-by: NeilBrown <neilb@suse.de>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>

Conflicts:

	block/blk-core.c
	drivers/md/dm.c
	drivers/md/raid1.c
	drivers/md/raid10.c
	drivers/staging/zram/zram_drv.c
	include/linux/blkdev.h

Change-Id: Ia945bf5891bf785723e75f8b558025132ed73935
Reviewed-on: http://coconut.acer.com.tw:8080/10908
Reviewed-by: vincent_tang <vincent_tang@acer.com.tw>
Tested-by: vincent_tang <vincent_tang@acer.com.tw>
Reviewed-by: elay_hu <elay_hu@acer.com.tw>
Reviewed-on: http://coconut.acer.com.tw:8080/11820