block: refactor generic_make_request
Move all the checks performed on a bio into a new helper, and call it as
soon as bio is submitted even if it is a re-submission from ->make_request.

We explicitly mark the new helper as beeing non-inlined as the stack
usage for printing the block device name in the failure case is quite
high and this a patch where we have to be extremely conservative about
stack usage.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Change-Id: Ia2391e994a31ba4049cbc330a71ce146f9d9e684
Reviewed-on: http://coconut.acer.com.tw:8080/10909
Reviewed-by: vincent_tang <vincent_tang@acer.com.tw>
Tested-by: vincent_tang <vincent_tang@acer.com.tw>
Reviewed-by: elay_hu <elay_hu@acer.com.tw>
Reviewed-on: http://coconut.acer.com.tw:8080/11821