block: Free queue resources at blk_release_queue()
A kernel crash is observed when a mounted ext3/ext4 filesystem is
physically removed. The problem is that blk_cleanup_queue() frees up
some resources eg by calling elevator_exit(), which are not checked for
in normal operation. So we should rather move these calls to the
destructor function blk_release_queue() as at that point all remaining
references are gone. However, in doing so we have to ensure that any
externally supplied queue_lock is disconnected as the driver might free
up the lock after the call of blk_cleanup_queue(),

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Jens Axboe <axboe@kernel.dk>

Conflicts:

	block/blk-sysfs.c

Change-Id: Ia09e7d02e5d030360b88deb0296668550dfc8fc0
Reviewed-on: http://coconut.acer.com.tw:8080/10911
Reviewed-by: vincent_tang <vincent_tang@acer.com.tw>
Tested-by: vincent_tang <vincent_tang@acer.com.tw>
Reviewed-by: elay_hu <elay_hu@acer.com.tw>
Reviewed-on: http://coconut.acer.com.tw:8080/11823